rename FilterByNameFilters to FilterArea
